+++ tests/test_api/test_methods/test_remove_chat_verification.py
@@ -1,0 +1,12 @@
+from aiogram.methods import RemoveChatVerification, VerifyChat
+from aiogram.types import Poll
+from tests.mocked_bot import MockedBot
+
+
+class TestRemoveChatVerification:
+    async def test_bot_method(self, bot: MockedBot):
+        prepare_result = bot.add_result_for(RemoveChatVerification, ok=True, result=True)
+
+        response: bool = await bot.remove_chat_verification(chat_id=42)
+        request = bot.get_request()
+        assert response == prepare_result.result